+++ test/collision_tests/decl_decl_collision_test.dart
@@ -36,6 +36,20 @@ void main() {
           rawType: 'int',
           rawValue: '0',
         ),
+
+        /// Conflicts across declarations.
+        Struc(name: 'testCrossDecl'),
+        Func(
+            name: 'testCrossDecl',
+            returnType: Type.nativeType(SupportedNativeType.Void)),
+        Constant(name: 'testCrossDecl', rawValue: '0', rawType: 'int'),
+        EnumClass(name: 'testCrossDecl'),
+
+        /// Conflicts with ffi library prefix, name of prefix is changed.
+        Struc(name: 'ffi'),
+        Func(
+            name: 'ffi_1',
+            returnType: Type.nativeType(SupportedNativeType.Void)),
       ]);
       final l2 = Library(name: 'Bindings', bindings: [
         Struc(name: 'TestStruc'),
@@ -62,6 +76,17 @@ void main() {
           rawType: 'int',
           rawValue: '0',
         ),
+        Struc(name: 'testCrossDecl', originalName: 'testCrossDecl'),
+        Func(
+            name: 'testCrossDecl_1',
+            originalName: 'testCrossDecl',
+            returnType: Type.nativeType(SupportedNativeType.Void)),
+        Constant(name: 'testCrossDecl_2', rawValue: '0', rawType: 'int'),
+        EnumClass(name: 'testCrossDecl_3'),
+        Struc(name: 'ffi'),
+        Func(
+            name: 'ffi_1',
+            returnType: Type.nativeType(SupportedNativeType.Void)),
       ]);
 
       expect(l1.generate(), l2.generate());